move synthesize method to model file
