+++ music/src/ddsp/ddsp.ts
@@ -14,13 +14,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import * as Tone from 'tone';
 import * as tf from '@tensorflow/tfjs';
-import { resampleAndMakeMono, hzToMidi } from '../core/audio_utils';
-import { mixAndJoinAudioData } from './audio_utils';
+import { hzToMidi } from '../core/audio_utils';
 import {
-  CROSSFADE_DURATION,
-  OUTPUT_SAMPLE_RATE,
   LOWEST_LD,
   CONF_SMOOTH_SIZE,
   CONF_THRESHOLD,
@@ -28,11 +24,9 @@ import {
   MIN_VRAM,
 } from './constants';
 import { AudioFeatures, ModelValues } from './interfaces';
-import { upsample_f0, shiftF0 } from '../spice/pitch_utils';
-import { Tensor3D, Tensor } from '@tensorflow/tfjs';
-import { addReverb } from './add_reverb';
-import { arrayBufferToAudioBuffer } from './buffer_utils';
-import { MODEL_SAMPLE_RATE, MODEL_FRAME_RATE } from '../spice/spice';
+import { shiftF0 } from '../spice/pitch_utils';
+import { Tensor3D } from '@tensorflow/tfjs';
+import { MODEL_FRAME_RATE } from '../spice/spice';
 
 async function memCheck() {
   const bytesPerMB = 1024 * 1024;
@@ -230,128 +224,6 @@ async function normalizeAudioFeatures(af: AudioFeatures, model: ModelValues) {
   return { f0_hz, loudness_db };
 }
 
-async function synthesize(
-  model: tf.GraphModel,
-  af: AudioFeatures,
-  settings?: ModelValues
-) {
-  const { f0_hz, loudness_db, confidences } = af;
-
-  // we need to upsample pitches because spice
-  // gives back something that is smaller
-  const upsampledPitches = upsample_f0(
-    f0_hz,
-    loudness_db.length,
-    settings.modelMaxFrameLength
-  );
-  const upsampledConfidences = upsample_f0(
-    confidences,
-    loudness_db.length,
-    settings.modelMaxFrameLength
-  );
-
-  const normalizedAudioFeatures = await normalizeAudioFeatures(
-    { f0_hz: upsampledPitches, loudness_db, confidences: upsampledConfidences },
-    settings
-  );
-
-  // we have to process the audio in chunks according
-  // to the model max frame length
-  // eg. if we have a 10s input and a 5s model, we'll have
-  // to splice our input into 2, process it twice, and then stitch it
-  // stitching the inputs require a linear mask to prevent an audible blip
-  // we overlap the last second and the first second of 2 conjoining chunks.
-  // eg. for a 10s input [1,2,3,4,5] + [5,6,7,8,9] + [9,10]
-  const audioChunks = [];
-  const inputFrameLength = normalizedAudioFeatures.loudness_db.length;
-  const inputAudioLengthInSecs = convertFrameToSecs(inputFrameLength);
-  let isLastChunks = false;
-  const usableAudioDuration = convertFrameToSecs(settings.modelMaxFrameLength);
-  let audioBuffer;
-
-  const mixBufferLength = CROSSFADE_DURATION * MODEL_SAMPLE_RATE;
-  for (
-    let i = 0;
-    i < inputAudioLengthInSecs;
-    i += usableAudioDuration - CROSSFADE_DURATION
-  ) {
-    const startFrameIndex = Math.floor(convertSecsToFrame(i));
-    const endFrameIndex = convertSecsToFrame(i + usableAudioDuration);
-    if (endFrameIndex > inputFrameLength) {
-      isLastChunks = true;
-    }
-
-    // tslint:disable-next-line: variable-name
-    const { f0_hz, loudness_db }: AudioFeatures = await resizeAudioFeatures(
-      normalizedAudioFeatures,
-      startFrameIndex,
-      endFrameIndex,
-      isLastChunks,
-      inputFrameLength
-    );
-
-    // tslint:disable-next-line: variable-name
-    const f0hzTensor = tf.tensor1d(f0_hz, 'float32');
-    // tslint:disable-next-line: variable-name
-    const loudnessTensor = tf.tensor1d(loudness_db, 'float32');
-
-    const result = await model.predict({
-      f0_hz: f0hzTensor,
-      loudness_db: loudnessTensor,
-    });
-
-    // apply a loudness modifier here to make some
-    // models louder/softer
-    const resultModified = (result as Tensor).mul(settings.postGain || 1);
-    const resultModifiedData = await resultModified.data();
-    audioChunks.push(resultModifiedData);
-
-    resultModified.dispose();
-    (result as Tensor).dispose();
-    f0hzTensor.dispose();
-    loudnessTensor.dispose();
-  }
-
-  // does not require stiching
-  if (inputAudioLengthInSecs <= usableAudioDuration) {
-    audioBuffer = audioChunks[0];
-  } else {
-    const buffers = [];
-    for (let i = 0; i < audioChunks.length; i++) {
-      const ac = audioChunks[i];
-      buffers.push(ac);
-    }
-    audioBuffer = mixAndJoinAudioData(buffers, mixBufferLength);
-  }
-
-  // trim off the extra bits
-  // tslint:disable-next-line: max-line-length
-  const trimmedAudioChannelData: Float32Array = (audioBuffer as Float32Array).slice(
-    0,
-    af.originalRecordedBufferLength
-  );
-
-  const audioCtx = new Tone.Context();
-  const trimmedAudioBuffer = arrayBufferToAudioBuffer(
-    audioCtx,
-    trimmedAudioChannelData,
-    MODEL_SAMPLE_RATE
-  );
-  const resampledAudio = await resampleAndMakeMono(
-    trimmedAudioBuffer,
-    OUTPUT_SAMPLE_RATE
-  );
-  const bufferWithReverbData = await addReverb({
-    audioCtx,
-    arrayBuffer: resampledAudio,
-    sampleRate: OUTPUT_SAMPLE_RATE,
-  });
-
-  model.dispose();
-
-  return bufferWithReverbData;
-}
-
 function convertSecsToFrame(secs: number) {
   return secs * MODEL_FRAME_RATE;
 }
@@ -359,4 +231,10 @@ function convertFrameToSecs(frameLength: number) {
   return frameLength / MODEL_FRAME_RATE;
 }
 
-export { memCheck, synthesize };
+export {
+  memCheck,
+  convertSecsToFrame,
+  convertFrameToSecs,
+  normalizeAudioFeatures,
+  resizeAudioFeatures,
+};
